+++ test/RMap.test.tsx
@@ -4,6 +4,7 @@ import {fireEvent, render} from '@testing-library/react';
 
 import {RMap, ROSM} from 'rlayers';
 import * as common from './common';
+import {RView} from 'rlayers/RMap';
 
 describe('<RMap>', () => {
     it('should display an OSM map', async () => {
@@ -25,12 +26,13 @@ describe('<RMap>', () => {
     it('should display an OSM map w/zoom', async () => {
         const map = React.createRef() as React.RefObject<RMap>;
         const {container} = render(
-            <RMap {...common.mapProps} ref={map} minZoom={1} maxZoom={10}>
+            <RMap {...common.mapProps} ref={map} minZoom={10} maxZoom={12}>
                 <ROSM />
             </RMap>
         );
-        expect(map.current?.ol.getView().getMinZoom()).toBe(1);
-        expect(map.current?.ol.getView().getMaxZoom()).toBe(10);
+        expect(map.current?.ol.getView().getMinZoom()).toBe(10);
+        expect(map.current?.ol.getView().getZoom()).toBe(11);
+        expect(map.current?.ol.getView().getMaxZoom()).toBe(12);
     });
     it('should display an OSM map w/resolution', async () => {
         const map = React.createRef() as React.RefObject<RMap>;
@@ -41,6 +43,16 @@ describe('<RMap>', () => {
         );
         expect(map.current?.ol.getView().getMinResolution()).toBe(1250);
         expect(map.current?.ol.getView().getMaxResolution()).toBe(10000);
+    });
+    it('should honor resolution when both zoom and resolution are specified', async () => {
+        const map = React.createRef() as React.RefObject<RMap>;
+        render(
+            <RMap initial={{...common.mapProps.initial, resolution: 2000}} ref={map}>
+                <ROSM />
+            </RMap>
+        );
+        expect(map.current?.ol.getView().getResolution()).toBe(2000);
+        expect(map.current?.ol.getView().getZoom()).not.toBe(11);
     });
     it('should handle Map events w/update', async () => {
         const mapEvents = [
@@ -82,7 +94,9 @@ describe('<RMap>', () => {
     });
     it('should support an external view state', async () => {
         const map = React.createRef() as React.RefObject<RMap>;
-        const mockSetView = jest.fn();
+        let view: RView = {center: [0, 0], zoom: 0};
+        const mockSetView = jest.fn((r) => (view = r));
+
         const {rerender} = render(
             <RMap {...common.mapProps} ref={map}>
                 <ROSM />
@@ -95,6 +109,19 @@ describe('<RMap>', () => {
         );
         map.current?.ol.dispatchEvent(common.createEvent('moveend', map.current.ol));
         expect(mockSetView).toHaveBeenCalledTimes(1);
+        expect(view.center).toEqual(expect.arrayContaining(common.mapProps.initial.center));
+        expect(view.zoom).toBe(common.mapProps.initial.zoom);
+        expect(view.resolution).toBeGreaterThan(0);
+
+        view.resolution = 2000;
+        rerender(
+            <RMap {...common.mapProps} ref={map} view={[view, mockSetView]}>
+                <ROSM />
+            </RMap>
+        );
+        expect(map.current?.ol.getView().getResolution()).toBe(2000);
+        expect(map.current?.ol.getView().getZoom()).not.toBe(11);
+
         rerender(
             <RMap {...common.mapProps} ref={map}>
                 <ROSM />